Added non-permanent entries to right prompt
Due to issues with resizing windows while using the powerline prompt,
I've decided to only add 'temporary' prompt items to the right-hand
side of the terminal prompt.

This seems like a good middle ground; Alacritty does not exhibit these
issues, but I can't copy and paste with a mouse on Alacritty, so I'm
back to Terminator (which freaked out every time I resized a window).