Added OpenShift Package Manager
This tooling allows user to build a local repository of OpenShift point
releases for custom tooling in the future. It seems like every version
of OpenShift in the wild is on a different version, and the versioning
is tightly coupled with both the cli tool and the kubectl tool. The real
benefit of this commit is having a repository of kubectls, since that's
the piece that will throw a fit if the versions are off by > 1.